refactor(lint): extract rewind and checkpoint helpers
Problem:
- CI repo standards rejected new growth in two oversized legacy files.

Root cause:
- Checkpoint load arbitration and rewind result plumbing remained embedded in their large coordinators.

Fix:
- Move checkpoint loading into a focused package file.
- Centralize rewind result normalization, target settling, and partial notice dispatch.
- Keep the existing file-size baseline unchanged.

Verification:
- go run ./tools/repolint
- go test ./internal/checkpoint -count=1
- go test -race ./internal/checkpoint -count=1
- pnpm lint:hooks
- pnpm typecheck
- pnpm test:typecheck
- rewind fork routing contract
